src/services/competition/requirements.txt
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
tomli==2.0.2
waitress==3.0.0
Werkzeug==3.0.4
coverage==7.6.1
psycopg2-binary==2.9.10
#psycopg2==2.9.10
requests==2.32.3